Structured existing code into multiple files
