+++ packages/buidler/contracts/Imports.sol
@@ -2,6 +2,8 @@ pragma solidity 0.4.24;
 
 import "@aragon/os/contracts/acl/ACL.sol";
 import "@aragon/os/contracts/kernel/Kernel.sol";
+import "@aragon/os/contracts/apm/APMRegistry.sol";
+import "@aragon/os/contracts/apm/Repo.sol";
 
 
 // HACK to workaround truffle artifact loading on dependencies
