+++ packages/buidler/tasks/task-names.ts
@@ -1,9 +1,6 @@
+// Aragon plugin tasks.
 export const TASK_START = 'start';
-export const TASK_START_NEW_DAO = 'start:new-dao';
-export const TASK_START_NEW_APP_PROXY = 'start:new-app-proxy';
-export const TASK_START_UPDATE_PROXY_IMPLEMENTATION =
-  'start:update-proxy-implementation';
 export const TASK_START_WATCH_CONTRACTS = 'start:watch-contracts';
 
-// buidler built-in
+// Buidler built-in tasks.
 export const TASK_COMPILE = 'compile';
